21.09.12【整理库存&库存转移】

+++ WcsFixPlatform/ViewModel/platform/track/OrganizeTrackViewModel.cs
@@ -257,7 +257,7 @@ namespace wcs.ViewModel
                     dtl_area_id = item.Area_Id,
                     DtlType = item.DtlType,
                     DtlStatus = item.DtlType == StockTransDtlTypeE.转移品种 ? StockTransDtlStatusE.整理中 : StockTransDtlStatusE.完成,
-                    dtl_level = (uint)item.Level,
+                    dtl_level = item.Level,
                 });
             }
