+++ module/rf/stock/RfStockTransDtl.cs
@@ -13,6 +13,6 @@ namespace module.rf
         public uint givetrackid { set; get; }// 卸货轨道
         public uint allqty { set; get; }// 全部库存数量
 
-        public uint level { set; get; }//品种等级
+        public byte level { set; get; }//品种等级
     }
 }
